modified formatters and report value to allow for float returns rather than force cast to long values

+++ source/Graphite.System/Kernel.cs
@@ -134,7 +134,7 @@ namespace Graphite.System
 
                 if (value.HasValue)
                 {
-                    channel.Report(config.Key, (long)value.Value);
+                    channel.Report(config.Key, value.Value);
                 }
             };
         }
